pyproject.toml
[tool.poetry]
name = "coffeehouse-softwaretools"
version = "0.1.0"
description = ""
authors = ["Alexandr Obolonkov <itbolonkov2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
django = "^4.1.7"
django-stubs = "^1.15.0"
pillow = "^9.4.0"
django-ckeditor = "^6.5.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"
pep8-naming = "^0.13.3"
mypy = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true